feat: rollup intial setup
