--- .npmignore
@@ -1,47 +1,0 @@
-.*.swp
-._*
-.DS_Store
-.git
-.gitlab-ci*
-.github
-.hg
-.npmrc
-.lock-wscript
-.svn
-.wafpickle-*
-.travis.yml
-.editorconfig
-.eslint*
-.yarnrc
-.babel*
-.stylelint*
-.storybook
-.yarn*
-.prettier*
-jest*
-postcss*
-prettier*
-rollup*
-vit*
-setupTests.*
-
-config.gypi
-CVS
-npm-debug.log
-
-yarn-debug.log*
-yarn-error.log*
-default.config
-rollup.config.js
-tsconfig.json
-
-docker*
-Docker*
-
-coverage
-config
-demo
-public
-src
-scripts
-storybook-static
