+++ package.json
@@ -12,10 +12,9 @@
   },
   "homepage": "https://tanishraj.github.io/barebones-ui/",
   "private": false,
-  "version": "0.1.1-beta.3",
-  "main": "dist/index.js",
-  "module": "dist/index.js",
-  "types": "dist/index.d.ts",
+  "version": "0.1.1-beta.4",
+  "main": "dist/cjs/index.js",
+  "module": "dist/esm/index.js",
   "type": "module",
   "files": [
     "dist",
@@ -50,9 +49,8 @@
     "@chromatic-com/storybook": "^3.2.2",
     "@eslint/js": "^9.13.0",
     "@rollup/plugin-commonjs": "^28.0.1",
-    "@rollup/plugin-json": "^6.1.0",
     "@rollup/plugin-node-resolve": "^15.3.0",
-    "@rollup/plugin-terser": "^0.4.4",
+    "@rollup/plugin-typescript": "^12.1.1",
     "@storybook/addon-a11y": "^8.4.4",
     "@storybook/addon-actions": "^8.4.4",
     "@storybook/addon-controls": "^8.4.4",
@@ -90,11 +88,11 @@
     "postcss": "^8.4.49",
     "prettier": "^3.3.3",
     "rollup": "^4.27.3",
-    "rollup-plugin-peer-deps-external": "^2.2.4",
+    "rollup-plugin-dts": "^6.1.1",
     "rollup-plugin-postcss": "^4.0.2",
-    "rollup-plugin-typescript2": "^0.36.0",
     "storybook": "^8.4.4",
     "tailwindcss": "^3.4.14",
+    "tslib": "^2.8.1",
     "typescript": "~5.6.2",
     "typescript-eslint": "^8.11.0",
     "vite": "^5.4.10",
