+++ rollup.config.js
@@ -1,70 +1,48 @@
-import peerDepsExternal from 'rollup-plugin-peer-deps-external';
 import resolve from '@rollup/plugin-node-resolve';
+import typescript from '@rollup/plugin-typescript';
 import commonjs from '@rollup/plugin-commonjs';
-import json from '@rollup/plugin-json';
-import typescript from 'rollup-plugin-typescript2';
+import dts from 'rollup-plugin-dts';
 import postcss from 'rollup-plugin-postcss';
-import { readFileSync } from 'fs';
-import typescriptEngine from 'typescript';
-import terser from '@rollup/plugin-terser';
+import packageJson from './package.json' with { type: 'json' };
 
-const packageJson = JSON.parse(readFileSync('./package.json'));
-
-export default {
-  input: './src/index.ts',
-  output: [
-    {
-      // file: packageJson.module,
-      // format: 'es',
-      // exports: 'named',
-      // sourcemap: false,
-      dir: 'dist',
-      format: 'es',
-      sourcemap: true,
-      preserveModules: true,
-    },
-  ],
-  plugins: [
-    peerDepsExternal(),
-    resolve({
-      extensions: ['.js', '.jsx', '.ts', '.tsx'],
-    }),
-    commonjs(),
-    json(),
-    typescript({
-      tsconfig: './tsconfig.app.json',
-      typescript: typescriptEngine,
-      sourceMap: false,
-      exclude: [
-        'coverage',
-        '.storybook',
-        'storybook-static',
-        'config',
-        'dist',
-        'node_modules/**',
-        '*.cjs',
-        '*.mjs',
-        '**/__snapshots__/*',
-        '**/__tests__',
-        '**/*.test.js+(|x)',
-        '**/*.test.ts+(|x)',
-        '**/*.mdx',
-        '**/*.story.ts+(|x)',
-        '**/*.story.js+(|x)',
-        '**/*.stories.ts+(|x)',
-        '**/*.stories.js+(|x)',
-        'setupTests.ts',
-        'vite.config.ts',
-        'vitest.config.ts',
-      ],
-    }),
-    postcss({
-      extract: true,
-      minimize: true,
-      sourceMap: true,
-      modules: false,
-    }),
-    terser(),
-  ],
-  external: ['react', 'react-dom'],
-};
+export default [
+  {
+    input: 'src/index.ts',
+    output: [
+      {
+        file: packageJson.main,
+        format: 'cjs',
+        sourcemap: true,
+      },
+      {
+        file: packageJson.module,
+        format: 'esm',
+        sourcemap: true,
+      },
+    ],
+    plugins: [
+      resolve({
+        extensions: ['.js', '.jsx', '.ts', '.tsx'],
+      }),
+      commonjs(),
+      typescript({
+        tsconfig: './tsconfig.app.json',
+        exclude: [
+          '**/*/test.tsx',
+          '**/*/test.ts',
+          '**/*/stories.tsx',
+          '**/*/stories.ts',
+        ],
+      }),
+      postcss({
+        extract: true,
+      }),
+    ],
+  },
+  {
+    input: 'dist/esm/types/index.d.ts',
+    output: [{ file: 'dist/index.d.ts', format: 'esm' }],
+    plugins: [dts()],
+    external: [/\.css$/],
+  },
+];
