+++ src/components/Button/Button.stories.tsx
@@ -1,6 +1,7 @@
 import { Meta, StoryFn } from '@storybook/react';
 
-import { Button, ButtonProps } from './Button';
+import { Button } from './Button';
+import { ButtonProps } from './types';
 
 export default {
   title: 'Components/Button',
