+++ src/components/Button/index.ts
@@ -1,2 +1,1 @@
-export { Button } from './Button';
-export type { ButtonProps } from './Button';
+export * from './Button';
