+++ src/components/Button/types.ts
@@ -1,0 +1,10 @@
+export type ButtonVariant = 'primary' | 'secondary' | 'ghost' | 'link';
+export type ButtonSize = 'sm' | 'md' | 'lg' | 'xl' | '2xl';
+
+export interface ButtonProps
+  extends React.ButtonHTMLAttributes<HTMLButtonElement> {
+  variant?: ButtonVariant;
+  size?: ButtonSize;
+  disabled?: boolean;
+  icon?: React.ReactNode;
+}
